+++ gDRplots/R/plot-pheatmap.R
@@ -495,7 +495,11 @@ pheatmap_with_anno_sa <- function(
   }
   
   # prep hm color palette
-  min_val <- min(t_mat_cvd[!is.infinite(t_mat_cvd)], na.rm = TRUE)
+  min_val <- ifelse(
+    all(is.infinite(t_mat_cvd) | is.na(t_mat_cvd)),
+    -1,
+    min(t_mat_cvd[!is.infinite(t_mat_cvd)], na.rm = TRUE)
+  )
   max_val <- ifelse(metric %in% c("x", "xc50", "x_max", "x_mean"), 1.0, max(t_mat_cvd, na.rm = TRUE))
   
   if (min_val == max_val) {