+++ gDRplots/tests/testthat/test-helpers-rmd.R
@@ -87,17 +87,12 @@ test_that("prep_nested_plot_chunk works as expected", {
 })
 
 test_that("escape_special_characters works as expected", {
-  expect_equal(escape_special_characters("ABC:123"), "ABC\\:123")
-  expect_equal(escape_special_characters("ABC:123", double_colon_escape = TRUE), "ABC\\\\:123")
+  expect_equal(escape_special_characters("ABC:123"), "ABC[colon]123")
   expect_equal(escape_special_characters("AD_12"), "AD_12")
   expect_equal(escape_special_characters("AD#12"), "AD[hash]12")
   expect_equal(escape_special_characters("AD/12"), "AD[slash]12")
   
   expect_error(escape_special_characters(123), "Assertion on 'x' failed: Must be of type 'string'")
-  expect_error(
-    escape_special_characters("ABC:123", double_colon_escape = "TRUE"),
-    "Assertion on 'double_colon_escape' failed: Must be of type 'logical flag'"
-  )
 })
 
 test_that("neutralize_spaces works as expected", {