+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+# gDRplots 0.0.65 - 2024-02-12
+* add support for chemical genomics analysis
+
 ## gDRplots 0.0.64 - 2025-02-12
 * downgrade reqs for Bioc pkgs
 
@@ -29,6 +32,9 @@
 
 ## gDRplots 0.0.56 - 2022-01-13
 * move ggrepel from Suggests to Imports
+
+## gDRplots 0.0.55 - 2025-01-03
+* add `get_r_file_path` function
 
 ## gDRplots 0.0.54 - 2024-12-20
 * refactor `prep_plot_chunk` to support simple nested lists of visualizations